requirements.txt
Django>=6.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
django-redis>=6.0
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=12.0
django-filter>=25.0
redis>=7.0
